backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.15.0
seaborn==0.12.2
gtts==2.3.2
pygame==2.5.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
